REFACTORING : socket connect refuse problem 클라이언트 연결 거부하는 상태
